Auto-update features and FBX export adjustments
Includes the basis for the auto-update functionality seen on other popular addons. It has yet to be fully implemented through the use of Github versions and/or tags. Adjusts the FBX export button to only export the selected armature and it's children regardless of how many there are in the blender scene.